Restructured the generateEndpoints util
