+++ resolvers/Videos/ShowResolver.js
@@ -1,7 +1,7 @@
 const axios = require("axios");
 const { filter, forEach } = require("lodash");
 
-const { generateVideoEndpoint } = require("../../utils/generateEndpoints");
+const generateVideoEndpoint = require("../../utils/generateEndpoints/Video");
 const setValue = require("../../utils/objects/setValue");
 
 const ShowVideoResolver = async (parent, args, context, info) => {
